do not include "search.html" in Authentfication filter
This resulted in a quite ugly and not easy to understand error message from the server. Since the search.html is able to detect by itself if logged in or not and redirects to the proper login page use this mechanism (even if you might see the search interface for a short time).

Security is not affected since all real data delivering servlets are still secured.