annis-widgets subproject is now compiling the GWT widget set
This means we will not check-in compiled GWT javascript any longer (which pollutes our GIT repository) but use the sub-project and every user has to compile it by itself once. Once we deploy annis-widgets to our repository this version can be used. You can also now clean the annis-gui projects safely without the need for the extreme long GWT recompiliation