removed the "scheme mode" mechanism and the "annopool" scheme
This was introduced to make it easier to test out new schemata and different approaches for querying. But it made it also more difficult to maintain and do unit testing. In theory new features needed to be included in every available scheme, not only the one activated by default. Since this becomes much too complex, the configuration possibility was removed. New schemes can be tried in separate branches. Since the benchmarking script of ANNIS has been stable for some time it should also be still possible to compare older versions.