refactoring (Default)AdministrationDao to be split into two parts and don't use an interface any longer
For support of transactions via annotations Spring uses proxy classes. Whenver you make a call inside the same class the @Transactional annotation is ignored. We have to make sure that the deletion of existing corpora at import is running inside it's own transaction (otherwise the "facts" table is blocked after dropping the child table). Thus we have to use a different class with it's own proxy for this task.