added Node Annotation Existence Validator functionality. This involves a number of changes in front and back ends. In the front end, I modified the QueryController class to perform validateQuery( ) with one more argument, the corpusName, which will be passed to back end at the time of validation (and obviously necessary for the new existence validator). Before, this argument was not passed from the front end and the corpusList in QueryServiceImpl was a empty linkedList (since previously no functionality in the validation needs the corpus name).
In the backend, I mainly added AnnotationExistenceValidator class. This is configured within the Java Spring framework, with the added entry in AnnisParser-Context.xml (including reference to annisDao). The new class simply get the available node annotations in the corpus (if selected), and checks if the nodes in the input query matches one of these valid annotation names.

One problem as of now is that, if no corpus is selected at the time that the validation happens, then the Node existence validation will not be performed. To change that situation we need to think of whether we want all validations to happen only when a corpus is selected by the user.