Attempt to fix #427.
In the tree set it's possible to have the same item but with different case. Why we don't want to have all upper case corpora first and then the lower case corpora we still can't use the case insensitive comparator. As a workaround the new CaseSensitiveOrder comparator is added which distinguishes between cases but which orders the items in a way that the same characters are actually near each other independently of the case.