don't re-use threads in PollControl to fix #433 and other similiar bugs
When doing background update jobs it is often necessary to call Vaadin.getSession() directly or indirectly (e.g. Helper.getAnnisWebResource() uses the session). This function is implemented using InheritableThreadLocal, so the sub-thread must be created in the right parent thread. This change will create a new thread from the parent thread by using a new SingleThreadExecutor on each call instead of a shared thread pool.