Refactor the error reporting for the different query functions and do expensive annotation name query validation only when actually executing the query.
There were serveral places doing the same kind of logic (read syntax errors, timeouts etc.). Now everything is handled by the QueryController#reportServiceException() function.

Also getting the list of annotation names for a corpus can be quite expensive for some corpora and that why it should not be done on the fly. The refactoring therefore won't do the expensive validation if the user types something but only when the query is actually executed somehow. Of course this also needs to be done in all different query functions.