Use list of corpus names in QueryData instead of their numerical ID.
Since the numerical ID is essential in several places of the legacy SQL generation, the complete SQL generation Spring config and several of the SQL generation classes have been removed. For some funtionality of the service which is not yet implemented in graphANNIS (e.g. "frequency" and list annotations), either an UnsupportedOperationException() is thrown or empty results are returned. The "matrix" function has been removed completly