Use the adapted NavigateableSinglePage class to remember the possible changed location and scroll position.
When the iframe is hidden and restored, the BrowserElement from Vaadin would just load the original URL. This is very inconvenient for the user, especially since the example page will put itself on the front quite often. The custom iframe component will remember the URL and scroll position.

For this to work, the help pages must be located on the same domain as the ANNIS interface, otherwise browser will disable the access.