Fix caching of meta annotations for the document browser.
The previous implementation did fetch meta annotations for all documents, even when not requested yet.
So when the cached map existed, we could be sure all documents are included.
The new implementation only fetches the meta data for the given document, thus we need to check if the document is part of the cache, too.