The context defined in the exporter panel was not taken into account.
We now synchronize the left/right context in the exporter panel and the one in the search options. Since this is doubled functionality, we might also consider to remove the left/right context in the exporter at some time. But having it synchronized seems less confusing than having two states.